requirements.txt
django-cms
#bootstrap-admin
django-compressor
django-filer
cmsplugin-filer
django-reversion
pillow
#-e git+https://github.com/fivethreeo/cmsplugin-blog.git#egg=cmsplugin-blog
django-blog-zinnia
cmsplugin_zinnia
zinnia-theme-bootstrap
django-app-namespace-template-loader
cmsplugin-contact
django-widget-tweaks
django-avatar
django-disqus
django-tinymce
